lc-dec: Improve the logic of biasing against motion modes
In parent version, OBMC, WARP motion modes were biased against by
increasing rd cost locally. Hence these motion modes were biased
against limitedly. In this CL, bias is introduced on rate,
distortion and sse members of RD_STATS structure after transform
search so that biased metrics are used in the calculation of rd
cost for subsequent rd decisions.

The Performance results of encoder and decoder for streams

      Decode Time                   BD-Rate Loss(%)
cpu  Reduction(%)  avg.psnr   ssim     vmaf   vmaf_neg  uvq_1p5
 1      0.693       0.0427   0.0740   0.0576   0.0633    0.5793
 2      0.619       0.0653   0.0754   0.0537   0.0419    0.0887
 3      0.127       0.0001  -0.0091  -0.0245  -0.0253    0.2953

STATS_CHANGED for low complexity decode

+++ av1/encoder/speed_features.h
@@ -1248,7 +1248,7 @@ typedef struct INTER_MODE_SPEED_FEATURES {
   // Percentage of scaling used to increase the rd cost of warp mode so that
   // encoder decisions are biased against local warp, favoring low complexity
   // modes.
-  int bias_warp_mode_rd_scale_pct;
+  float bias_warp_mode_rd_scale_pct;
 
   // Percentage of scaling used to increase the rd cost of obmc motion mode so
   // that encoder decisions are biased against local obmc, favoring low
